+++ nekoyume/Assets/_Scripts/UI/Model/CollectionModel.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 49825c1de45949cba748e19b08dd0cc1
+timeCreated: 1707211369
